[chore] 중복 파일 삭제(UITextField+)
